Move some methods to Service from Controller.
