+++ src/main/scala/gitbucket/core/controller/RepositoryViewerController.scala
@@ -7,14 +7,13 @@ import gitbucket.core.plugin.PluginRegistry
 import gitbucket.core.repo.html
 import gitbucket.core.helper
 import gitbucket.core.service._
+import gitbucket.core.service.RepositoryCommitFileService.CommitFile
 import gitbucket.core.util._
-import gitbucket.core.util.JGitUtil._
 import gitbucket.core.util.StringUtil._
 import gitbucket.core.util.SyntaxSugars._
 import gitbucket.core.util.Implicits._
 import gitbucket.core.util.Directory._
 import gitbucket.core.model.{Account, CommitState, CommitStatus, WebHook}
-import gitbucket.core.service.WebHookService._
 import gitbucket.core.view
 import gitbucket.core.view.helpers
 import org.apache.commons.compress.archivers.{ArchiveEntry, ArchiveOutputStream}
@@ -24,15 +23,12 @@ import org.apache.commons.compress.compressors.bzip2.BZip2CompressorOutputStream
 import org.apache.commons.compress.compressors.gzip.GzipCompressorOutputStream
 import org.apache.commons.compress.compressors.xz.XZCompressorOutputStream
 import org.apache.commons.compress.utils.IOUtils
-import org.scalatra.forms._
 import org.apache.commons.io.FileUtils
-import org.ec4j.core.model.PropertyType
+import org.scalatra.forms._
 import org.eclipse.jgit.api.{ArchiveCommand, Git}
 import org.eclipse.jgit.archive.{TgzFormat, ZipFormat}
-import org.eclipse.jgit.dircache.{DirCache, DirCacheBuilder}
 import org.eclipse.jgit.errors.MissingObjectException
 import org.eclipse.jgit.lib._
-import org.eclipse.jgit.transport.{ReceiveCommand, ReceivePack}
 import org.eclipse.jgit.treewalk.TreeWalk
 import org.eclipse.jgit.treewalk.filter.PathFilter
 import org.json4s.jackson.Serialization
@@ -42,6 +38,7 @@ import org.scalatra.i18n.Messages
 class RepositoryViewerController
     extends RepositoryViewerControllerBase
     with RepositoryService
+    with RepositoryCommitFileService
     with AccountService
     with ActivityService
     with IssuesService
@@ -64,6 +61,7 @@ class RepositoryViewerController
  */
 trait RepositoryViewerControllerBase extends ControllerBase {
   self: RepositoryService
+    with RepositoryCommitFileService
     with AccountService
     with ActivityService
     with IssuesService
@@ -319,13 +317,34 @@ trait RepositoryViewerControllerBase extends ControllerBase {
       CommitFile(line.substring(0, i).trim, line.substring(i + 1).trim)
     }
 
+    val newFiles = files.map { file =>
+      file.copy(name = if (form.path.length == 0) file.name else s"${form.path}/${file.name}")
+    }
+
     commitFiles(
       repository = repository,
       branch = form.branch,
       path = form.path,
       files = files,
-      message = form.message.getOrElse("Add files via upload")
-    )
+      message = form.message.getOrElse("Add files via upload"),
+      loginAccount = context.loginAccount.get
+    ) {
+      case (git, headTip, builder, inserter) =>
+        JGitUtil.processTree(git, headTip) { (path, tree) =>
+          if (!newFiles.exists(_.name.contains(path))) {
+            builder.add(JGitUtil.createDirCacheEntry(path, tree.getEntryFileMode, tree.getEntryObjectId))
+          }
+        }
+
+        newFiles.foreach { file =>
+          val bytes =
+            FileUtils.readFileToByteArray(new File(getTemporaryDir(session.getId), FileUtil.checkFilename(file.id)))
+          builder.add(
+            JGitUtil.createDirCacheEntry(file.name, FileMode.REGULAR_FILE, inserter.insert(Constants.OBJ_BLOB, bytes))
+          )
+          builder.finish()
+        }
+    }
 
     if (form.path.length == 0) {
       redirect(s"/${repository.owner}/${repository.name}/tree/${form.branch}")
@@ -394,7 +413,8 @@ trait RepositoryViewerControllerBase extends ControllerBase {
       content = appendNewLine(convertLineSeparator(form.content, form.lineSeparator), form.lineSeparator),
       charset = form.charset,
       message = form.message.getOrElse(s"Create ${form.newFileName}"),
-      commit = form.commit
+      commit = form.commit,
+      loginAccount = context.loginAccount.get
     )
 
     redirect(
@@ -417,7 +437,8 @@ trait RepositoryViewerControllerBase extends ControllerBase {
       } else {
         form.message.getOrElse(s"Rename ${form.oldFileName.get} to ${form.newFileName}")
       },
-      commit = form.commit
+      commit = form.commit,
+      loginAccount = context.loginAccount.get
     )
 
     redirect(
@@ -436,7 +457,8 @@ trait RepositoryViewerControllerBase extends ControllerBase {
       content = "",
       charset = "",
       message = form.message.getOrElse(s"Delete ${form.fileName}"),
-      commit = form.commit
+      commit = form.commit,
+      loginAccount = context.loginAccount.get
     )
 
     println(form.path)
@@ -928,185 +950,6 @@ trait RepositoryViewerControllerBase extends ControllerBase {
 
   case class UploadFiles(branch: String, path: String, fileIds: Map[String, String], message: String) {
     lazy val isValid: Boolean = fileIds.nonEmpty
-  }
-
-  case class CommitFile(id: String, name: String)
-
-  private def commitFiles(
-    repository: RepositoryService.RepositoryInfo,
-    files: Seq[CommitFile],
-    branch: String,
-    path: String,
-    message: String
-  ) = {
-    // prepend path to the filename
-    val newFiles = files.map { file =>
-      file.copy(name = if (path.length == 0) file.name else s"${path}/${file.name}")
-    }
-
-    _commitFile(repository, branch, message) {
-      case (git, headTip, builder, inserter) =>
-        JGitUtil.processTree(git, headTip) { (path, tree) =>
-          if (!newFiles.exists(_.name.contains(path))) {
-            builder.add(JGitUtil.createDirCacheEntry(path, tree.getEntryFileMode, tree.getEntryObjectId))
-          }
-        }
-
-        newFiles.foreach { file =>
-          val bytes =
-            FileUtils.readFileToByteArray(new File(getTemporaryDir(session.getId), FileUtil.checkFilename(file.id)))
-          builder.add(
-            JGitUtil.createDirCacheEntry(file.name, FileMode.REGULAR_FILE, inserter.insert(Constants.OBJ_BLOB, bytes))
-          )
-          builder.finish()
-        }
-    }
-  }
-
-  private def commitFile(
-    repository: RepositoryService.RepositoryInfo,
-    branch: String,
-    path: String,
-    newFileName: Option[String],
-    oldFileName: Option[String],
-    content: String,
-    charset: String,
-    message: String,
-    commit: String
-  ) = {
-
-    val newPath = newFileName.map { newFileName =>
-      if (path.length == 0) newFileName else s"${path}/${newFileName}"
-    }
-    val oldPath = oldFileName.map { oldFileName =>
-      if (path.length == 0) oldFileName else s"${path}/${oldFileName}"
-    }
-
-    _commitFile(repository, branch, message) {
-      case (git, headTip, builder, inserter) =>
-        if (headTip.getName == commit) {
-          val permission = JGitUtil
-            .processTree(git, headTip) { (path, tree) =>
-              // Add all entries except the editing file
-              if (!newPath.contains(path) && !oldPath.contains(path)) {
-                builder.add(JGitUtil.createDirCacheEntry(path, tree.getEntryFileMode, tree.getEntryObjectId))
-              }
-              // Retrieve permission if file exists to keep it
-              oldPath.collect { case x if x == path => tree.getEntryFileMode.getBits }
-            }
-            .flatten
-            .headOption
-
-          newPath.foreach { newPath =>
-            builder.add(JGitUtil.createDirCacheEntry(newPath, permission.map { bits =>
-              FileMode.fromBits(bits)
-            } getOrElse FileMode.REGULAR_FILE, inserter.insert(Constants.OBJ_BLOB, content.getBytes(charset))))
-          }
-          builder.finish()
-        }
-    }
-  }
-
-  private def _commitFile(repository: RepositoryService.RepositoryInfo, branch: String, message: String)(
-    f: (Git, ObjectId, DirCacheBuilder, ObjectInserter) => Unit
-  ) = {
-
-    LockUtil.lock(s"${repository.owner}/${repository.name}") {
-      using(Git.open(getRepositoryDir(repository.owner, repository.name))) { git =>
-        val loginAccount = context.loginAccount.get
-        val builder = DirCache.newInCore.builder()
-        val inserter = git.getRepository.newObjectInserter()
-        val headName = s"refs/heads/${branch}"
-        val headTip = git.getRepository.resolve(headName)
-
-        f(git, headTip, builder, inserter)
-
-        val commitId = JGitUtil.createNewCommit(
-          git,
-          inserter,
-          headTip,
-          builder.getDirCache.writeTree(inserter),
-          headName,
-          loginAccount.fullName,
-          loginAccount.mailAddress,
-          message
-        )
-
-        inserter.flush()
-        inserter.close()
-
-        val receivePack = new ReceivePack(git.getRepository)
-        val receiveCommand = new ReceiveCommand(headTip, commitId, headName)
-
-        // call post commit hook
-        val error = PluginRegistry().getReceiveHooks.flatMap { hook =>
-          hook.preReceive(repository.owner, repository.name, receivePack, receiveCommand, loginAccount.userName)
-        }.headOption
-
-        error match {
-          case Some(error) =>
-            // commit is rejected
-            // TODO Notify commit failure to edited user
-            val refUpdate = git.getRepository.updateRef(headName)
-            refUpdate.setNewObjectId(headTip)
-            refUpdate.setForceUpdate(true)
-            refUpdate.update()
-
-          case None =>
-            // update refs
-            val refUpdate = git.getRepository.updateRef(headName)
-            refUpdate.setNewObjectId(commitId)
-            refUpdate.setForceUpdate(false)
-            refUpdate.setRefLogIdent(new PersonIdent(loginAccount.fullName, loginAccount.mailAddress))
-            refUpdate.update()
-
-            // update pull request
-            updatePullRequests(repository.owner, repository.name, branch)
-
-            // record activity
-            val commitInfo = new CommitInfo(JGitUtil.getRevCommitFromId(git, commitId))
-            recordPushActivity(repository.owner, repository.name, loginAccount.userName, branch, List(commitInfo))
-
-            // create issue comment by commit message
-            createIssueComment(repository.owner, repository.name, commitInfo)
-
-            // close issue by commit message
-            if (branch == repository.repository.defaultBranch) {
-              closeIssuesFromMessage(message, loginAccount.userName, repository.owner, repository.name).foreach {
-                issueId =>
-                  getIssue(repository.owner, repository.name, issueId.toString).foreach { issue =>
-                    callIssuesWebHook("closed", repository, issue, baseUrl, loginAccount)
-                    PluginRegistry().getIssueHooks
-                      .foreach(_.closedByCommitComment(issue, repository, message, loginAccount))
-                  }
-              }
-            }
-
-            // call post commit hook
-            PluginRegistry().getReceiveHooks.foreach { hook =>
-              hook.postReceive(repository.owner, repository.name, receivePack, receiveCommand, loginAccount.userName)
-            }
-
-            //call web hook
-            callPullRequestWebHookByRequestBranch("synchronize", repository, branch, context.baseUrl, loginAccount)
-            val commit = new JGitUtil.CommitInfo(JGitUtil.getRevCommitFromId(git, commitId))
-            callWebHookOf(repository.owner, repository.name, WebHook.Push) {
-              getAccountByUserName(repository.owner).map { ownerAccount =>
-                WebHookPushPayload(
-                  git,
-                  loginAccount,
-                  headName,
-                  repository,
-                  List(commit),
-                  ownerAccount,
-                  oldId = headTip,
-                  newId = commitId
-                )
-              }
-            }
-        }
-      }
-    }
   }
 
   private val readmeFiles = PluginRegistry().renderableExtensions.map { extension =>
